Move keyboard factory to other file
