+++ Shared/Sources/Canvas/A11yControlLayer.swift
@@ -78,3 +78,4 @@ public extension A11yControlLayer {
         return control
     }
 }
+
